Moved CatogoryX interface to the group-item-variants-generator.
