+++ libs/platform/core/src/orchestration/injection/injection.ts
@@ -1,4 +1,3 @@
-import { AppPluginBeforeApply } from '@spryker-oryx/core';
 import {
   createInjector,
   destroyInjector,
@@ -6,7 +5,7 @@ import {
   InjectorOptions,
   Provider,
 } from '@spryker-oryx/di';
-import { App, AppPlugin, AppRef } from '../app';
+import {App, AppPlugin, AppPluginBeforeApply, AppRef} from '../app';
 
 export const InjectionPluginName = 'core$injection';
